Move contrib normalization operators to a separate file
Move the `contrib` submodule of `src/ops/norm.rs` to
`src/ops/norm/contrib.rs` and move the tests for the contrib operators
into a `tests` module in that file.
